#565 | refactor(masonry): move mock palette config from mock to mocks
