+++ modules/masonry/src/mocks/palette.ts
